+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.1.1 under development
 
-- no changes in this release.
+- Enh #98: Add ability to set default theme for concrete widget (@vjik)
 
 ## 2.1.0 November 16, 2023